update : code refactoring and countCharacter function test
