+++ index.html
@@ -134,6 +134,7 @@
   </main>
   <script src="./theme.js"></script>
   <script src="./app.js"></script>
+  <script src="./util.js"></script>
 </body>
 
 </html>
